+++ rs/pocket_ic_server/tests/test.rs
@@ -422,6 +422,36 @@ fn test_specified_id() {
 }
 
 #[test]
+fn test_dashboard() {
+    let (server_url, _) = start_server_helper(None, Some(5), true);
+    let subnet_config_set = SubnetConfigSet {
+        nns: true,
+        application: 1,
+        ..Default::default()
+    };
+    let pic = PocketIc::from_config_and_server_url(subnet_config_set, server_url.clone());
+
+    // retrieve the NNS and application subnets
+    let topology = pic.topology();
+    let nns_subnet = topology.get_nns().unwrap();
+    let app_subnet = topology.get_app_subnets()[0];
+
+    let canister_1 = pic.create_canister_on_subnet(None, None, nns_subnet);
+    assert_eq!(pic.get_subnet(canister_1).unwrap(), nns_subnet);
+    let canister_2 = pic.create_canister_on_subnet(None, None, app_subnet);
+    assert_eq!(pic.get_subnet(canister_2).unwrap(), app_subnet);
+
+    let client = Client::new();
+    let dashboard_url = format!("{}instances/{}/_/dashboard", server_url, pic.instance_id());
+    let dashboard = client.get(dashboard_url).send().unwrap();
+    let page = String::from_utf8(dashboard.bytes().unwrap().to_vec()).unwrap();
+    assert!(page.contains(&canister_1.to_string()));
+    assert!(page.contains(&canister_2.to_string()));
+    assert!(page.contains(&nns_subnet.to_string()));
+    assert!(page.contains(&app_subnet.to_string()));
+}
+
+#[test]
 fn pocket_ic_server_binary_name() {
     let bin_path = std::env::var_os("POCKET_IC_BIN").expect("Missing PocketIC binary");
     let new_bin_path = format!("{}_", bin_path.to_str().unwrap());